+++ simulation-system/python/csle-common/csle_common/dao/datasets/traces_dataset.py
@@ -7,7 +7,7 @@ class TracesDataset:
     """
 
     def __init__(self, name: str, description: str, download_count: int, file_path: str, url: str, date_added,
-                 num_traces: int, num_attributes_per_time_step: int, size_in_gb: int, compressed_size_in_gb: int,
+                 num_traces: int, num_attributes_per_time_step: int, size_in_gb: float, compressed_size_in_gb: float,
                  citation: str, num_files: int, data_schema: str, file_format: str):
         """
         Initializes the DTO
